apps/fastapi/pyproject.toml
[project]
name = "fastapi-epitope-prediction"
version = "0.1.0"
description = "FastAPI for B-cell and T-cell epitope prediction."
authors = [
    {name = "trevorpfiz", email = "elektrikspark@gmail.com"},
    {name = "zacharypfiz", email = "ztpfizcode@gmail.com"},
]
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "aiofiles>=24.1.0",
    "boto3>=1.35.29",
    "fastapi[standard]>=0.115.0",
    "huggingface-hub>=0.25.1",
    "mangum>=0.19.0",
    "openai>=1.50.2",
    "pandas>=2.2.3",
    "pydantic-settings>=2.5.2",
    "python-jose[cryptography]>=3.3.0",
    "requests>=2.32.3",
    "supabase>=2.8.1",
]

[tool.uv]
dev-dependencies = [
    "pylint-pydantic>=0.3.2",
    "pytest>=8.3.3",
    "python-dotenv>=1.0.1",
    "ruff>=0.6.8",
]
